+++ src/Utils/utility_functions.jl
@@ -1,3 +1,5 @@
+using Dates
+
 """
 $(TYPEDSIGNATURES)
 Check whether elements of a vector `v` are strictly increasing."""